renames PopoverContent to PopoverBody and PopoverTitle to PopoverHeader
